+++ lab7/src/cpe200/BinaryCalculator.java
@@ -2,14 +2,7 @@ package cpe200;
 
 import java.math.BigDecimal;
 
-public class BinaryCalculator {
-    private BigDecimal firstOperand;
-    private BigDecimal secondOperand;
-
-    public BinaryCalculator() {
-        firstOperand = new BigDecimal(0);
-        secondOperand = new BigDecimal(0);
-    }
+public class BinaryCalculator extends  BaseCalculator {
 
     public void setFirstOperand(IOperand operand) {
         check(operand);
@@ -22,44 +15,23 @@ public class BinaryCalculator {
     }
 
     public String add() throws RuntimeException {
-        if(getSecondOperand().intValue()<0||getFirstOperand().intValue()<0){
-            throw new RuntimeException("Negative value");
-        }
         return Integer.toBinaryString(getFirstOperand().add(getSecondOperand()).stripTrailingZeros().intValue());
     }
 
     public String subtract() throws RuntimeException {
-        if(getSecondOperand().intValue()<0||getFirstOperand().intValue()<0){
-            throw new RuntimeException("Negative value");
-        }
         return Integer.toBinaryString(getFirstOperand().subtract(getSecondOperand()).stripTrailingZeros().intValue());
     }
 
     public String multiply() throws RuntimeException {
-        if(getSecondOperand().intValue()<0||getFirstOperand().intValue()<0){
-            throw new RuntimeException("Negative value");
-        }
         return Integer.toBinaryString(getFirstOperand().multiply(getSecondOperand()).stripTrailingZeros().intValue());
     }
 
     /* This method should throw an exception when divide by zero */
     public String division() throws RuntimeException {
-        if(getSecondOperand().intValue()<0||getFirstOperand().intValue()<0){
-            throw new RuntimeException("Negative value");
-        }
-        secondOperand.toString();
-        if(secondOperand.equals("0")) {
-            throw new ArithmeticException("Divide by zero");
-        }
-        else{
-            return Integer.toBinaryString(getFirstOperand().divide(getSecondOperand(), 5, BigDecimal.ROUND_HALF_UP).stripTrailingZeros().intValue());
-        }
+        return Integer.toBinaryString(getFirstOperand().divide(getSecondOperand(), 5, BigDecimal.ROUND_HALF_UP).stripTrailingZeros().intValue());
     }
 
     public String power() throws RuntimeException {
-        if(getSecondOperand().intValue()<0||getFirstOperand().intValue()<0){
-            throw new RuntimeException("Negative value");
-        }
         return  Integer.toBinaryString(getFirstOperand().pow(getSecondOperand().intValue()).stripTrailingZeros().intValue());
     }
     private void check(IOperand checkbi) throws ArithmeticException {
@@ -67,12 +39,6 @@ public class BinaryCalculator {
         else throw new ArithmeticException();
     }
 
-    public  BigDecimal getFirstOperand(){
-        return firstOperand;
-    }
-    public  BigDecimal getSecondOperand(){
-        return secondOperand;
-    }
     public void setFirstOperand(BigDecimal firstOperand) {
         this.firstOperand = firstOperand;
     }
